refactor: separate matrices implementation, and modify headers and using aliases style
