+++ include/keisan/matrix/matrix.hpp
@@ -21,22 +21,14 @@
 #ifndef KEISAN__MATRIX__MATRIX_HPP_
 #define KEISAN__MATRIX__MATRIX_HPP_
 
-#include <gtest/gtest.h>
-
-#include <algorithm>
 #include <ostream>
 
 namespace keisan
 {
 
-template<size_t M, size_t N>
-class Matrix;
-
+// Forward declaration
 template<size_t N>
 class Vector;
-
-template<size_t M, size_t N>
-inline std::ostream & operator<<(std::ostream & out, const Matrix<M, N> & matrix);
 
 template<size_t M, size_t N>
 class Matrix
@@ -86,251 +78,11 @@ private:
   double data[M * N];
 };
 
-template<size_t M, size_t N>
-std::ostream & operator<<(std::ostream & out, const Matrix<M, N> & matrix)
-{
-  out << "[";
-  for (size_t i = 0; i < M; ++i) {
-    if (i > 0) {
-      out << ",";
-    }
-
-    out << "[";
-    for (size_t j = 0; j < N; ++j) {
-      if (j > 0) {
-        out << ",";
-      }
-
-      out << matrix[i][j];
-    }
-
-    out << "]";
-  }
-
-  out << "]";
-
-  return out;
-}
-
-template<size_t M, size_t N>
-Matrix<M, N>::Matrix()
-{
-}
-
-template<size_t M, size_t N>
-template<typename ... Types>
-Matrix<M, N>::Matrix(const double & value, Types ... the_rest)
-: data{value, the_rest ...}
-{
-}
-
-template<size_t M, size_t N>
-Matrix<M, N>::Matrix(const Matrix<M, N> & matrix)
-{
-  *this = matrix;
-}
-
-template<size_t M, size_t N>
-Matrix<M, N> Matrix<M, N>::zero()
-{
-  Matrix<M, N> matrix;
-  for (size_t i = 0; i < M; ++i) {
-    for (size_t j = 0; j < N; ++j) {
-      matrix[i][j] = 0.0;
-    }
-  }
-
-  return matrix;
-}
-
-template<size_t M, size_t N>
-Matrix<M, N> & Matrix<M, N>::operator=(const Matrix<M, N> & matrix)
-{
-  std::copy(matrix.data, matrix.data + M * N, data);
-
-  return *this;
-}
-
-template<size_t M, size_t N>
-bool Matrix<M, N>::operator==(const Matrix<M, N> & matrix) const
-{
-  using testing::internal::FloatingPoint;
-  for (size_t i = 0; i < M * N; ++i) {
-    if (!FloatingPoint<double>(data[i]).AlmostEquals(FloatingPoint<double>(matrix.data[i]))) {
-      return false;
-    }
-  }
-
-  return true;
-}
-
-template<size_t M, size_t N>
-bool Matrix<M, N>::operator!=(const Matrix<M, N> & matrix) const
-{
-  return !(*this == matrix);
-}
-
-template<size_t M, size_t N>
-Matrix<M, N> & Matrix<M, N>::operator+=(const Matrix<M, N> & matrix)
-{
-  for (size_t i = 0; i < M * N; ++i) {
-    data[i] += matrix.data[i];
-  }
-
-  return *this;
-}
-
-template<size_t M, size_t N>
-Matrix<M, N> & Matrix<M, N>::operator-=(const Matrix<M, N> & matrix)
-{
-  for (size_t i = 0; i < M * N; ++i) {
-    data[i] -= matrix.data[i];
-  }
-
-  return *this;
-}
-
-template<size_t M, size_t N>
-Matrix<M, N> & Matrix<M, N>::operator+=(const double & value)
-{
-  for (size_t i = 0; i < M * N; ++i) {
-    data[i] += value;
-  }
-
-  return *this;
-}
-
-template<size_t M, size_t N>
-Matrix<M, N> & Matrix<M, N>::operator-=(const double & value)
-{
-  for (size_t i = 0; i < M * N; ++i) {
-    data[i] -= value;
-  }
-
-  return *this;
-}
-
-template<size_t M, size_t N>
-Matrix<M, N> & Matrix<M, N>::operator*=(const double & value)
-{
-  for (size_t i = 0; i < M * N; ++i) {
-    data[i] *= value;
-  }
-
-  return *this;
-}
-
-template<size_t M, size_t N>
-Matrix<M, N> & Matrix<M, N>::operator/=(const double & value)
-{
-  for (size_t i = 0; i < M * N; ++i) {
-    data[i] /= value;
-  }
-
-  return *this;
-}
-
-
-template<size_t M, size_t N>
-template<size_t O>
-Matrix<M, O> Matrix<M, N>::operator*(const Matrix<N, O> & matrix)
-{
-  Matrix<M, O> new_matrix;
-  for (size_t i = 0; i < M; ++i) {
-    for (size_t j = 0; j < O; ++j) {
-      new_matrix[i][j] = 0.0;
-      for (size_t k = 0; k < N; ++k) {
-        new_matrix[i][j] += (*this)[i][k] * matrix[k][j];
-      }
-    }
-  }
-
-  return new_matrix;
-}
-
-template<size_t M, size_t N>
-Vector<M> Matrix<M, N>::operator*(const Vector<N> & vector)
-{
-  return (Vector<M>) operator*((Matrix<N, 1>& )vector);
-}
-
-template<size_t M, size_t N>
-Matrix<M, N> Matrix<M, N>::operator+(const Matrix<M, N> & matrix) const
-{
-  auto new_matrix = *this;
-  new_matrix += matrix;
-
-  return new_matrix;
-}
-
-template<size_t M, size_t N>
-Matrix<M, N> Matrix<M, N>::operator-(const Matrix<M, N> & matrix) const
-{
-  auto new_matrix = *this;
-  new_matrix -= matrix;
-
-  return new_matrix;
-}
-
-template<size_t M, size_t N>
-Matrix<M, N> Matrix<M, N>::operator+(const double & value) const
-{
-  auto new_matrix = *this;
-  new_matrix += value;
-
-  return new_matrix;
-}
-
-template<size_t M, size_t N>
-Matrix<M, N> Matrix<M, N>::operator-(const double & value) const
-{
-  auto new_matrix = *this;
-  new_matrix -= value;
-
-  return new_matrix;
-}
-
-template<size_t M, size_t N>
-Matrix<M, N> Matrix<M, N>::operator*(const double & value) const
-{
-  auto new_matrix = *this;
-  new_matrix *= value;
-
-  return new_matrix;
-}
-
-template<size_t M, size_t N>
-Matrix<M, N> Matrix<M, N>::operator/(const double & value) const
-{
-  auto new_matrix = *this;
-  new_matrix /= value;
-
-  return new_matrix;
-}
-
-template<size_t M, size_t N>
-Matrix<M, N> Matrix<M, N>::operator-() const
-{
-  Matrix<M, N> matrix;
-  for (size_t i = 0; i < M * N; ++i) {
-    matrix.data[i] = -data[i];
-  }
-
-  return matrix;
-}
-
-template<size_t M, size_t N>
-double * Matrix<M, N>::operator[](size_t pos)
-{
-  return data + (pos * N);
-}
-
-template<size_t M, size_t N>
-const double * Matrix<M, N>::operator[](size_t pos) const
-{
-  return data + (pos * N);
-}
-
 }  // namespace keisan
+
+template<size_t M, size_t N>
+inline std::ostream & operator<<(std::ostream & out, const keisan::Matrix<M, N> & matrix);
+
+#include "keisan/matrix/matrix.impl.hpp"
 
 #endif  // KEISAN__MATRIX__MATRIX_HPP_
